* xfer-src/dest-null.c: fix Solaris 8 compilation problem: isprint()   expects an int, not a char.
git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@895 a8d146d6-cc15-0410-8900-af154a0219e0

+++ xfer-src/dest-null.c
@@ -85,7 +85,7 @@ read_data_thread(gpointer data)
 	    ssize_t i;
 	    g_printf("data: ");
 	    for (i = 0; i < len; i++) {
-		if (isprint(buf[i]))
+		if (isprint((int)buf[i]))
 		    g_printf("%c", buf[i]);
 		else
 		    g_printf("\\x%02x", (int)buf[i]);
